cronjober/requirements.txt
httpx
# onvif-py3
onvif2-zeep